refactor: Move formik instantiation into useSchema hook
